Unicode: Hebrew examples
